% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{listFiles}
\alias{listFiles}
\title{List files}
\usage{
listFiles(path = "")
}
\arguments{
\item{path}{The path of directory.}
}
\value{
A list of available files and metadata, including path, type, size.
}
\description{
List files available on the portal.
}
\examples{
\donttest{
listFiles("data")
}
}
